@@ -31,12 +31,14 @@ import (
 	"github.com/docker/docker/api/types/network"
 	"github.com/docker/docker/api/types/volume"
 	"github.com/docker/docker/client"
+	"github.com/hashicorp/go-version"
 	"github.com/pkg/errors"
 	"github.com/rs/zerolog/log"
 	"github.com/valyala/fasthttp"
 
 	"pmm-dump/internal/test/util"
 	grafanaClient "pmm-dump/pkg/grafana/client"
+	pkgUtil "pmm-dump/pkg/util"
 )
 
 const (
@@ -219,6 +221,16 @@ func (pmm *PMM) DontCleanup() {
 	pmm.dontCleanup = true
 }
 
+// Returns major version.
+func (pmm *PMM) GetVersion() *version.Version {
+	v1, err := version.NewVersion(pmm.pmmVersion)
+	if err != nil {
+		panic(fmt.Sprintf("cannot get version: %v", err))
+	} else {
+		return v1
+	}
+}
+
 func (pmm *PMM) SetVersion(version string) {
 	pmm.pmmVersion = version
 }
@@ -336,6 +348,7 @@ func (pmm *PMM) deploy(ctx context.Context) error {
 		return errors.Wrap(err, "failed to add mongo to PMM")
 	}
 
+	pmm.Log("Ping clickhouse with driver")
 	tCtx, cancel = context.WithTimeout(ctx, execTimeout)
 	defer cancel()
 	if err := util.RetryOnError(tCtx, func() error {
@@ -365,8 +378,14 @@ func (pmm *PMM) Restart(ctx context.Context) error {
 
 	tCtx, cancel := context.WithTimeout(ctx, getTimeout)
 	defer cancel()
-	if err := getUntilOk(tCtx, pmm.PMMURL()+"/v1/server/version"); err != nil && !errors.Is(err, io.EOF) {
-		return errors.Wrap(err, "failed to ping PMM")
+	if pkgUtil.CheckIsVer2(pmm.GetVersion()) {
+		if err := getUntilOk(tCtx, pmm.PMMURL()+"/v1/version"); err != nil && !errors.Is(err, io.EOF) {
+			return errors.Wrap(err, "failed to ping PMM")
+		}
+	} else {
+		if err := getUntilOk(tCtx, pmm.PMMURL()+"/v1/server/version"); err != nil && !errors.Is(err, io.EOF) {
+			return errors.Wrap(err, "failed to ping PMM")
+		}
 	}
 	return nil
 }
