@@ -18,7 +18,6 @@ import (
 	"archive/tar"
 	"bytes"
 	"compress/gzip"
-	"context"
 	"crypto/rand"
 	"fmt"
 	"io"
@@ -30,7 +29,7 @@ import (
 )
 
 func TestImport(t *testing.T) {
-	ctx := context.Background()
+	ctx := t.Context()
 
 	tests := []struct {
 		name          string
@@ -226,8 +225,7 @@ func writeFakeFile(t *testing.T, w io.Writer, opts fakeFileOpts) {
 			t.Fatal(err, "failed to write chunk content")
 		}
 	}
-
-	for i := 0; i < 10; i++ {
+	for i := range 10 {
 		var content bytes.Buffer
 		switch {
 		case opts.withEmptyChunk && i == 5:
