@@ -103,10 +103,10 @@ func TestWriteChunk(t *testing.T) {
 			if err != nil {
 				t.Fatal(err)
 			}
-			s := NewSource(grafanaC, Config{
+			s := NewSource(grafanaC, &Config{
 				ConnectionURL: server.URL,
 				NativeData:    tt.nativeData,
-				ContentLimit:  uint64(tt.contentLimit),
+				ContentLimit:  tt.contentLimit,
 			})
 
 			data, err := generateFakeChunk(tt.metricsSize)
@@ -139,7 +139,7 @@ func generateFakeChunk(size int) ([]byte, error) {
 		return nil, errors.Wrap(err, "marshal metrics")
 	}
 	var data []byte
-	for i := 0; i < size; i++ {
+	for range size {
 		data = append(data, metricsData...)
 	}
 	return compressData(data)
